requirements.txt
pandas==2.0.3
numpy
scikit-learn
matplotlib
seaborn
sentence-transformers
umap-learn
hdbscan
openpyxl